+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - [#122](https://github.com/Zuoqiu-Yingyi/siyuan-theme-dark-plus/issues/122) 修复悬浮工具栏无法在标题栏拖拽问题 | Fixed the issue that the floating toolbar could not be dragged in the title bar.
 - 修复快捷键设置界面列表辅助线样式 | Fixed the list guide style of shortcut key setting interface.
 - 保存悬浮工具栏的位置与状态 | Save the position and state of the floating toolbar.
+- 修复悬浮工具栏折叠展开时的位置问题 | Fixed the position issue when the floating toolbar was collapsed and expanded.
 
 ## v1.3.5/2022-10-17